+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [x.x.x] - 2025-xx-xx
 
+
 ### Fixed
 
 -   `pixelator single-cell demux` respects the `--cores` option.
+
+### Added
+
+-   Validation on antibody panel sequence lengths and characters.
+-   If panel contains uniprot_id column, it will be added to adata.var.
 
 ## [0.21.4] - 2025-08-27